# api_docs/openapi.yaml
openapi: 3.0.2
info:
  title: Announcements API
  version: 0.1.0
paths:
  /announcements:
    get:
      tags:
        - Announcements
      summary: Get Announcements
      operationId: get_announcements_announcements_get
      parameters:
        - required: false
          schema:
            title: Limit
            maximum: 100
            minimum: 1
            type: integer
            default: 10
          name: limit
          in: query
        - required: false
          schema:
            title: Skip
            minimum: 0
            type: integer
            default: 0
          name: skip
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Announcements Announcements Get
                type: array
                items:
                  $ref: '#/components/schemas/Announcement'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Announcements
      summary: Create Announcement
      operationId: create_announcement_announcements_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '409':
          description: AlreadyExists Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        
components:
  schemas:
    Announcement:
      title: Announcement
      required:
        - title
      type: object
      properties:
        title:
          title: Title
          type: string
        date:
          title: Date
          type: string
          format: date-time
        description:
          title: Description
          type: string
          default: ''
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string